+++ camera_tracking/include/camera_tracking/ball_detection.h
