+++ camera_tracking/src/ball_detection.cpp
@@ -24,7 +24,7 @@
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#include "camera_tracking/BallDetection.h"
+#include "camera_tracking/ball_detection_node.h"
 
 
 int main(int argc, char *argv[])
